+++ packages/adapter-cloudflare/index.js
@@ -14,10 +14,17 @@ export default function (options = {}) {
 
 			builder.rimraf(dest);
 			builder.rimraf(tmp);
+
+			builder.mkdirp(dest);
 			builder.mkdirp(tmp);
 
-			// generate 404.html first which can then be overridden by prerendering, if the user defined such a page
-			await builder.generateFallback(path.join(dest, '404.html'));
+			// generate plaintext 404.html first which can then be overridden by prerendering, if the user defined such a page
+			const fallback = path.join(dest, '404.html');
+			if (options.fallback === 'spa') {
+				await builder.generateFallback(fallback);
+			} else {
+				writeFileSync(fallback, 'Not Found');
+			}
 
 			const dest_dir = `${dest}${builder.config.kit.paths.base}`;
 			const written_files = builder.writeClient(dest_dir);
@@ -28,7 +35,8 @@ export default function (options = {}) {
 			writeFileSync(
 				`${tmp}/manifest.js`,
 				`export const manifest = ${builder.generateManifest({ relativePath })};\n\n` +
-					`export const prerendered = new Set(${JSON.stringify(builder.prerendered.paths)});\n`
+					`export const prerendered = new Set(${JSON.stringify(builder.prerendered.paths)});\n\n` +
+					`export const app_path = ${JSON.stringify(builder.getAppPath())};\n`
 			);
 
 			writeFileSync(