+++ packages/kit/src/exports/vite/index.js
@@ -128,9 +128,15 @@ const warning_preprocessor = {
  * @param {string} dependency
  */
 async function resolve_peer_dependency(dependency) {
-	// @ts-expect-error the types are wrong
-	const resolved = await imr.resolve(dependency, pathToFileURL(process.cwd() + '/dummy.js'));
-	return import(resolved);
+	try {
+		// @ts-expect-error the types are wrong
+		const resolved = await imr.resolve(dependency, pathToFileURL(process.cwd() + '/dummy.js'));
+		return import(resolved);
+	} catch (e) {
+		throw new Error(
+			`Could not resolve peer dependency "${dependency}" relative to your project — please install it and try again.`
+		);
+	}
 }
 
 /**
@@ -216,6 +222,7 @@ async function kit({ svelte_config }) {
 	let initial_config;
 
 	const service_worker_entry_file = resolve_entry(kit.files.serviceWorker);
+	const parsed_service_worker = path.parse(kit.files.serviceWorker);
 
 	const sourcemapIgnoreList = /** @param {string} relative_path */ (relative_path) =>
 		relative_path.includes('node_modules') || relative_path.includes(kit.outDir);
@@ -355,7 +362,24 @@ async function kit({ svelte_config }) {
 	const plugin_virtual_modules = {
 		name: 'vite-plugin-sveltekit-virtual-modules',
 
-		async resolveId(id) {
+		async resolveId(id, importer) {
+			// If importing from a service-worker, only allow $service-worker & $env/static/public, but none of the other virtual modules.
+			// This check won't catch transitive imports, but it will warn when the import comes from a service-worker directly.
+			// Transitive imports will be caught during the build.
+			if (importer) {
+				const parsed_importer = path.parse(importer);
+
+				const importer_is_service_worker =
+					parsed_importer.dir === parsed_service_worker.dir &&
+					parsed_importer.name === parsed_service_worker.name;
+
+				if (importer_is_service_worker && id !== '$service-worker' && id !== '$env/static/public') {
+					throw new Error(
+						`Cannot import ${id} into service-worker code. Only the modules $service-worker and $env/static/public are available in service workers.`
+					);
+				}
+			}
+
 			// treat $env/static/[public|private] as virtual
 			if (id.startsWith('$env/') || id.startsWith('__sveltekit/') || id === '$service-worker') {
 				return `\0virtual:${id}`;
@@ -400,7 +424,7 @@ async function kit({ svelte_config }) {
 				case env_dynamic_public:
 					// populate `$env/dynamic/public` from `window`
 					if (browser) {
-						return `export const env = ${global}.env ?? (await import(/* @vite-ignore */ ${global}.base + '/' + '${kit.appDir}/env.js')).env;`;
+						return `export const env = ${global}.env;`;
 					}
 
 					return create_dynamic_module(