+++ packages/kit/src/runtime/server/ambient.d.ts
@@ -1,8 +1,4 @@
 declare module '__SERVER__/internal.js' {
 	export const options: import('types').SSROptions;
-	export const get_hooks: () => Promise<{
-		handle?: import('@sveltejs/kit').Handle;
-		handleError?: import('@sveltejs/kit').HandleServerError;
-		handleFetch?: import('@sveltejs/kit').HandleFetch;
-	}>;
+	export const get_hooks: () => Promise<Partial<import('types').ServerHooks>>;
 }