+++ packages/kit/test/build-errors/env.spec.js
@@ -49,3 +49,27 @@ test('$env/static/private is not dynamically importable from the client', () => 
 		/.*Cannot import \$env\/static\/private into client-side code:.*/gs
 	);
 });
+
+test('$env/dynamic/private is not importable from the service worker', () => {
+	assert.throws(
+		() =>
+			execSync('pnpm build', {
+				cwd: path.join(process.cwd(), 'apps/service-worker-private-env'),
+				stdio: 'pipe',
+				timeout: 60000
+			}),
+		/.*Cannot import \$env\/dynamic\/private into service-worker code.*/gs
+	);
+});
+
+test('$env/dynamic/public is not importable from the service worker', () => {
+	assert.throws(
+		() =>
+			execSync('pnpm build', {
+				cwd: path.join(process.cwd(), 'apps/service-worker-dynamic-public-env'),
+				stdio: 'pipe',
+				timeout: 60000
+			}),
+		/.*Cannot import \$env\/dynamic\/public into service-worker code.*/gs
+	);
+});